+++ contracts/reflect/tests/integration.rs
@@ -95,7 +95,7 @@ fn reflect() {
         CustomMsg::Debug("Hi, Dad!".to_string()).into(),
         StakingMsg::Delegate {
             validator: HumanAddr::from("validator"),
-            amount: coin(100, "stake"),
+            amount: coin(100, "ustake"),
         }
         .into(),
     ];